Init commit of this branch. changing the structure of the project
