--- Student/backend/index.js
@@ -1,41 +1,0 @@
-const express = require("express");
-const mongoose = require("mongoose");
-const bodyParser = require("body-parser");
-const passport = require("passport");
-const users = require("./routes/api/users");
-const results = require("./routes/api/results");
-
-const app = express();
-const cors = require('cors')
-
-//CORS middleware
-app.use(cors())
-
-// Bodyparser middleware
-app.use(
-    bodyParser.urlencoded({
-        extended: false
-    })
-);
-app.use(bodyParser.json());
-
-// DB Config
-mongoose.set('strictQuery', false);
-const db = require("./config/keys").mongoURI;
-
-// Connect to MongoDB
-mongoose.connect(db, { useNewUrlParser: true })
-    .then(() => console.log("MongoDB successfully connected"))
-    .catch(err => console.log(err));
-
-// Passport middleware
-app.use(passport.initialize());
-// Passport config
-require("./config/passport")(passport);
-
-// Routes
-app.use("/api/users", users);
-app.use('/api/results', results);
-
-const port = process.env.PORT || 3001;
-app.listen(port, () => console.log(`Server up and running on port ${port} !`));
