--- Student/backend/models/User.js
@@ -1,22 +1,0 @@
-const mongoose = require('mongoose')
-const Schema = mongoose.Schema;
-// Create Schema
-const UserSchema = new Schema({
-  name: {
-    type: String,
-    required: true
-  },
-  email: {
-    type: String,
-    required: true
-  },
-  password: {
-    type: String,
-    required: true
-  },
-  date: {
-    type: Date,
-    default: Date.now
-  }
-});
-module.exports = User = mongoose.model("student-credentials", UserSchema);
