--- Student/backend/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "backend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "start": "nodemon index.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "body-parser": "^1.20.1",
-    "canvas": "^2.11.0",
-    "child_process": "^1.0.2",
-    "concurrently": "^7.6.0",
-    "cors": "^2.8.5",
-    "dotenv": "^16.0.3",
-    "express": "^4.18.2",
-    "face-api.js": "^0.22.2",
-    "http-proxy-middleware": "^2.0.6",
-    "is-empty": "^1.2.0",
-    "jsonwebtoken": "^9.0.0",
-    "mongoose": "^6.9.0",
-    "nodemon": "^2.0.21",
-    "passport": "^0.6.0",
-    "passport-jwt": "^4.0.1",
-    "validator": "^13.7.0"
-  }
-}
