--- Teacher/backend/routes/api/results.js
@@ -1,11 +1,0 @@
-const express = require("express");
-const router = express.Router();
-
-const {DisplayAll} = require('../../controllers/resultController')
-
-// @route POST api/results/scores
-// @desc displays results of quiz
-// @access Public
-router.get('/scores', DisplayAll);
-
-module.exports = router
