+++ server/.gitignore
@@ -1,3 +1,2 @@
 /node_modules
-package-lock.json
 /config/keys.js
