+++ server/routes/student/users.js
@@ -1,4 +1,5 @@
-const {SignIn, Register} = require('../../controllers/userController')
+// const {SignIn, Register} = 
+const {SignIn, Register } = require('../../controllers/Student/userController')
 const express = require("express");
 const router = express.Router();
 // Load input validation
